requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
linecache2==1.0.0
openai==1.46.0
pydantic==2.9.2
pydantic-core==2.23.4
pypdf2==3.0.1
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
traceback2==1.4.0
typing-extensions==4.12.2
unittest2==1.1.0
